Refactor String interpolator code and document limitations
